tr2/objects: split worker module
