--- src/tr2/game/objects/creatures/worker.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void Worker1_Setup(void);
-void Worker2_Setup(void);
-void Worker3_Setup(void);
-void Worker4_Setup(void);
-void Worker5_Setup(void);
