+++ src/tr2/game/objects/creatures/worker_1.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void Worker1_Setup(void);
