+++ src/tr2/game/objects/creatures/worker_2.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void Worker2_Setup(void);
+void Worker5_Setup(void);
