+++ src/tr2/game/objects/creatures/worker_3.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void Worker3_Setup(void);
+void Worker4_Setup(void);
