+++ src/tr2/game/objects/creatures/worker_common.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "global/const.h"
+
+#define WORKER_RADIUS (WALL_L / 10) // = 102
