+++ src/tr2/game/objects/setup.c
@@ -27,7 +27,9 @@
 #include "game/objects/creatures/tiger.h"
 #include "game/objects/creatures/trex.h"
 #include "game/objects/creatures/winston.h"
-#include "game/objects/creatures/worker.h"
+#include "game/objects/creatures/worker_1.h"
+#include "game/objects/creatures/worker_2.h"
+#include "game/objects/creatures/worker_3.h"
 #include "game/objects/creatures/xian_knight.h"
 #include "game/objects/creatures/xian_spearman.h"
 #include "game/objects/creatures/yeti.h"
