+++ src/tr2/meson.build
@@ -157,7 +157,9 @@ dll_sources = [
   'game/objects/creatures/tiger.c',
   'game/objects/creatures/trex.c',
   'game/objects/creatures/winston.c',
-  'game/objects/creatures/worker.c',
+  'game/objects/creatures/worker_1.c',
+  'game/objects/creatures/worker_2.c',
+  'game/objects/creatures/worker_3.c',
   'game/objects/creatures/xian_knight.c',
   'game/objects/creatures/xian_spearman.c',
   'game/objects/creatures/yeti.c',
